MdeModulePkg: Add GetMemoryMapEx for unaccepted memory support
From: Dionna Glaze <dionnaglaze@google.com>

This is a proposed specification change to add a new boot service
call, GetMemoryMapEx, as suggested by Gerd Hoffman on the Qemu
mailing list.

The addition of this boot service allows the guest OS to tell
the UEFI whether or not it supports a particular memory map feature
before the UEFI passes a map that it won't understand.

In this case, we have only one feature that we'd like to determine
the support for: unaccepted memory. This allows SEV-SNP to use
the memory type, since "SEV-SNP support" in an OS does not require
supporting this particular memory type.

CC: Thomas Lendacky <thomas.lendacky@amd.com>
CC: Gerd Hoffman <kravel@redhat.com>
CC: Jiewen Yao <jiewen.yao@intel.com>
CC: Jiaqi Gao <jiaqi.gao@intel.com>
Signed-off-by: Dionna Glaze <dionnaglaze@google.com>