Add OsIndication for unaccepted memory support
The addition of unaccepted memory types introduces a problem: if the
UEFI creates an unaccepted memory region, but the guest OS doesn't
support them, then the guest is unable to access most of the system
memory they have allotted.

Instead of leaving memory unaccepted by default, accept memory if the OS
has not indicated that it supports unaccepted memory.

Signed-off-by: Dionna Glaze <dionnaglaze@google.com>